Move utility functions to utilities folder
